+++ nativeui/table.h
@@ -49,8 +49,8 @@ class NATIVEUI_EXPORT Table : public View {
   void AddColumnWithOptions(const std::string& title,
                             const ColumnOptions& options);
   int GetColumnCount() const;
-  void SetRowHeight(int height);
-  int GetRowHeight() const;
+  void SetRowHeight(float height);
+  float GetRowHeight() const;
   void SelectRow(int row);
   int GetSelectedRow() const;
